+++ base_de_datos/modelos/EDIC_registros_rech.js
@@ -9,7 +9,7 @@ module.exports = (sequelize, dt) => {
 		valor_aceptado: {type: dt.STRING(50)},
 
 		motivo_id: {type: dt.INTEGER},
-		duracion: {type: dt.INTEGER},
+		duracion: {type: dt.DECIMAL},
 
 		input_por_id: {type: dt.INTEGER},
 		input_en: {type: dt.DATE},
